Them cap nhat xoa danh muc

+++ lib/ui/brand/brand_list_tile.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:myshop/ui/brand/brand_manager.dart';
+import 'package:myshop/ui/brand/edit_brand_screen.dart';
+import 'package:myshop/ui/shared/dialog_utils.dart';
 import 'package:provider/provider.dart';
 import '../../models/brand.dart';
 
@@ -12,24 +14,42 @@ class BrandListTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ListTile(
-      title: Text(brand.title, style: TextStyle(color: Colors.black)),
-      subtitle: Text(brand.content),
-      trailing: SizedBox(
-        width: 100,
-        child: Row(children: <Widget>[
-          buildEditButton(context),
-          buildDeleteButton(context),
-        ]),
+    return Padding(
+      padding: const EdgeInsets.all(8.0),
+      child: Container(
+        decoration: BoxDecoration(
+          border: Border.all(
+            color: Color.fromARGB(255, 66, 65, 65),
+            style: BorderStyle.solid,
+            width: 1.0,
+          ),
+          color: Colors.transparent,
+          borderRadius: BorderRadius.circular(8.0),
+        ),
+        child: ListTile(
+          title: Text(brand.content, style: TextStyle(color: Colors.black)),
+          subtitle: Text('\$${brand.denominations}'),
+          trailing: SizedBox(
+            width: 100,
+            child: Row(children: <Widget>[
+              buildEditButton(context),
+              buildDeleteButton(context),
+            ]),
+          ),
+        ),
       ),
     );
   }
 
   Widget buildEditButton(BuildContext context) {
+    // print(brand.id);
     return IconButton(
       icon: const Icon(Icons.edit),
       onPressed: () {
-        print('edit brand');
+        Navigator.of(context).pushNamed(
+          EditBrandScreen.routeName,
+          arguments: brand.id,
+        );
       },
       color: Theme.of(context).primaryColor,
     );
@@ -37,15 +57,21 @@ class BrandListTile extends StatelessWidget {
 
   Widget buildDeleteButton(BuildContext context) {
     return IconButton(
-        onPressed: () {
-          // print('delete brand');
-          context.read<BrandManager>().deleteBrand(brand.id!);
-          ScaffoldMessenger.of(context)
-            ..hideCurrentSnackBar()
-            ..showSnackBar(const SnackBar(content: Text(
-              'Brand deleted',
-              textAlign: TextAlign.center,
+        onPressed: () async {
+          final shouldDelete = await showConfirmDialog(
+              context, "Bạn muốn xóa danh mục đang chọn ?", "Xóa danh mục");
+
+          // print(shouldDelete);
+          if (shouldDelete!) {
+            context.read<BrandManager>().deleteBrand(brand.id!);
+            ScaffoldMessenger.of(context)
+              ..hideCurrentSnackBar()
+              ..showSnackBar(const SnackBar(
+                  content: Text(
+                'Danh mục đã được xóa',
+                textAlign: TextAlign.center,
               )));
+          }
         },
         icon: const Icon(Icons.delete),
         color: Theme.of(context).colorScheme.error);
